pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = ""
authors = [
    {name = "GuiBarao",email = "guibarao04@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standart] (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "firebird-driver (>=2.0.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "selenium (>=4.34.2,<5.0.0)",
    "validate-docbr (>=1.11.1,<2.0.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
]

[tool.poetry]
packages = [{include = "myapp", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
